requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
PyJWT==2.10.1
cryptography==45.0.6
